Fix randint error. (Python 3.12+)
Corrected an issue that was caused by calculating a random integer from a range of floats, which is not allowed in Python 3.12 and later. Several values that were previously defined as floats have been re-defined as integers, and several other values, such as the HUD size, have been adjusted accordingly.